SimulatedAnnealingABC

Documentation for SimulatedAnnealingABC.

SimulatedAnnealingABC.SABCresult

SimulatedAnnealingABC.SABCstate

Base.show

SimulatedAnnealingABC.build_cdf

SimulatedAnnealingABC.build_cdf

SimulatedAnnealingABC.estimate_jump_covariance

SimulatedAnnealingABC.initialization

SimulatedAnnealingABC.proposal

SimulatedAnnealingABC.proposal

SimulatedAnnealingABC.resample_population

SimulatedAnnealingABC.sabc

SimulatedAnnealingABC.update_epsilon

SimulatedAnnealingABC.SABCresult — Type

Holds results

population: vector of parameter samples from the approximate posterior

u: transformed distances

ρ: distances

state: state of algorithm

SimulatedAnnealingABC.SABCstate — Type

Holds state of algorithm

Base.show — Method

Function for pretty printing

SimulatedAnnealingABC.build_cdf — Method

Estimate the empirical cdf of data x smoothed by interpolation.

Returns a function.

SimulatedAnnealingABC.build_cdf — Method

Construct the empirical cdf for each summary statistics

It assumes every row x corresponds to a sample.

Returns a function, that applies the corresponding cdf to each statistics.

SimulatedAnnealingABC.estimate_jump_covariance — Method

Estimate the covariance for the jump distributions from a population

SimulatedAnnealingABC.initialization — Method

Initialisation step

Arguments

See docs for sabc

Return

An object of type SABCresult

SimulatedAnnealingABC.proposal — Method

Proposal for n-dimensions, n > 1

SimulatedAnnealingABC.proposal — Method

Proposal for 1-dimension

SimulatedAnnealingABC.resample_population — Method

Resample population

SimulatedAnnealingABC.sabc — Method

sabc(f_dist::Function, prior::Distribution, args...;
      n_particles = 100, n_simulation = 10_000,
      type = :single,
      resample = 2*n_particles,
      v=1.0, β=0.8, δ=0.1,
      checkpoint_history = 1,
      show_progressbar::Bool = !is_logging(stderr),
      show_checkpoint = is_logging(stderr) ? 100 : Inf,
      kwargs...)

Simulated Annealing Approximate Bayesian Inference Algorithm

Arguments

f_dist: Function that one or more distances between data and a random sample from the likelihood. The first argument must be the parameter vector.

prior: A Distribution defining the prior.

args...: Further arguments passed to f_dist

n_particles: Desired number of particles.

n_simulation: maximal number of simulations from f_dist.

v = 1.0: Tuning parameter for XXX

β = 0.8: Tuning parameter for XXX

δ = 0.1: Tuning parameter for XXX

type = :single -> single-ϵ = :multi -> multi-ϵ = :hybrid -> hybrid multi-u-single-ϵ

resample: After how many accepted population updates?

checkpoint_history = 1: every how many population updates distances and epsilons are stored

show_progressbar::Bool = !is_logging(stderr): defaults to true for interactive use.

show_checkpoint::Int = 100: every how many population updates algorithm state is displayed. By default disabled for for interactive use.

kwargs...: Further arguments passed to f_dist`

Return

An object of type SABCresult

SimulatedAnnealingABC.update_epsilon — Method

Update ϵ Single-epsilon: see eq(31) in Albert et al., Statistics and Computing 25, 2015 Multi-epsilon: new update rule