SimulatedAnnealingABC

Documentation for SimulatedAnnealingABC.

This package provides different SimulatedAnnealingABC (SABC) algorithms for Approximate Bayesian Computation (ABC) (sometimes also called simulation-based inference or likelihood-free inference).

Note

Can you evaluate the density of your posterior? Then you should most likely not be using this or any other ABC package! A traditional MCMC algorithm will be much more efficient.

Usage

Getting started

A minimal example how to use this package for inference:

# Define a stochastic model.
# Your real model should be so complex, that it would be too
# complicated to compute it's likelihood function.
function my_stochastic_model(θ, n)
    rand(Normal(θ[1], θ[2]), n)
end

# define prior of the parameters
prior = product_distribution([Normal(0,2),   # theta[1]
                              Uniform(0,2)]) # theta[2]

# Simulate some observation data
y_obs = rand(Normal(2, 0.5), 10)

# Define a function that simulate with my_stochastic_model and then
# measure the distances of the simulated and the observed data with
# two summary statistics
function f_dist(θ; y_obs)
    y_sim = my_stochastic_model(θ, length(y_obs))

    (
        abs(mean(y_obs) - mean(y_sim)),
        abs(sum(abs2, y_obs) - sum(abs2, y_sim))
    )
end

## Sample Posterior
res = sabc(f_dist, prior;
           n_particles = 1000, n_simulation = 100_000, y_obs=y_obs)

## Improve the results by running the inference for longer
res2 = update_population!(res, f_dist, prior;
                          n_simulation = 50_000, y_obs=y_obs)

Logging and progress bar

There are two ways sabc can inform about the progress of the ongoing inference run:

Progress bar. By default it is shown in interactive sessions. It can be disabled with the argument (show_progressbar = false)

Logging statements. The argument show_checkpoint controls how often a summary of the current state is logged. For long running computations in a cluster environment this is more convenient than the progress bar.

You can set the logging level to Warn to suppress all logging statements:

using Logging
global_logger(ConsoleLogger(stderr, Logging.Warn))

... run sabc() ...

How to disable multi-threading

TODO

Worked example

Translate and update example from readme.

explain pro and cons of the different algo types.

API

SimulatedAnnealingABC.sabc — Function

sabc(f_dist::Function, prior::Distribution, args...;
      n_particles = 100, n_simulation = 10_000,
      type = :single,
      resample = 2*n_particles,
      v=1.0, β=0.8, δ=0.1,
      checkpoint_history = 1,
      show_progressbar::Bool = !is_logging(stderr),
      show_checkpoint = is_logging(stderr) ? 100 : Inf,
      kwargs...)

Simulated Annealing Approximate Bayesian Inference Algorithm

Arguments

f_dist: Function that one or more distances between data and a random sample from the likelihood. The first argument must be the parameter vector.

prior: A Distribution defining the prior.

args...: Further arguments passed to f_dist

n_particles: Desired number of particles.

n_simulation: maximal number of simulations from f_dist.

v = 1.0: Tuning parameter for XXX

β = 0.8: Tuning parameter for XXX

δ = 0.1: Tuning parameter for XXX

type = :single: Choose algorithm, either :single ,:multi, or :hybrid

resample: After how many accepted population updates?

checkpoint_history = 1: every how many population updates distances and epsilons are stored

show_progressbar::Bool = !is_logging(stderr): defaults to true for interactive use.

show_checkpoint::Int = 100: every how many population updates algorithm state is displayed. By default disabled for for interactive use.

kwargs...: Further arguments passed to f_dist`

Return

An object of type SABCresult

SimulatedAnnealingABC.update_population! — Function

update_population!(population_state::SABCresult,
                   f_dist, prior, args...;
                   n_simulation,
                   v=1.0, β=0.8, δ=0.1,
                   resample = 2*length(population_state.population),
                   checkpoint_history = 1,
                   show_progressbar::Bool = !is_logging(stderr),
                   show_checkpoint = is_logging(stderr) ? 100 : Inf,
                   kwargs...)

Updates particles with n_simulation and applies importance sampling if needed. Modifies population_state.

Arguments

See docstring for sabc.

SimulatedAnnealingABC.SABCresult — Type

Holds results from a SABC run with fields:

population: vector of parameter samples from the approximate posterior

u: transformed distances

ρ: distances

state: state of algorithm

The history of ϵ can be accessed with the field state.ϵ_history.

Related Julia Packages

Approximate Bayesian Computation

ABCdeZ.jl - Approximate Bayesian Computation (ABC) with differential evolution (de) moves and model evidence (Z) estimates.

ApproxBayes.jl - Implements basic ABC rejection sampler and sequential monte carlo algorithm (ABC SMC) as in Toni. et al 2009 as well as model selection versions of both (Toni. et al 2010).

GpABC.jl - ABC with emulation on Gaussian Process Regression.

KissABC.jl - Implementation of Multiple Affine Invariant Sampling for efficient Approximate Bayesian Computation. Is looking for a new maintainer.

Misc

SimulationBasedInference.jl Despite the name this package seems to focus on traditional Bayesian inference methods and not ABC, i.e. it assumes we can evaluate the density of the posterior.

References

Albert, C., Künsch, H.R., Scheidegger, A., 2015. A simulated annealing approach to approximate Bayes computations. Statistics and computing 25, 1217–1232. https://doi.org/10.1007/s11222-014-9507-8