SimulatedAnnealingABC

Documentation for SimulatedAnnealingABC.

SimulatedAnnealingABC.initialization_noninf

SimulatedAnnealingABC.rho_mean

SimulatedAnnealingABC.sabc_noninf

SimulatedAnnealingABC.schedule

SimulatedAnnealingABC.initialization_noninf — Method

Initialisation step

Arguments

f_dist: Function that returns a the distance between data and a random sample from the likelhood. The first argument must be the parameter vector.

d_prior: Function that returns the density of the prior distribution.

r_prior: Function which returns one vector of a random realization of the prior distribution.

n_sample: Desired number of samples.

eps_init: Initial epsilon.

iter_max: Maximal number of iterations.

v: Tuning parameter.

beta: Tuning parameter.

args...: Further arguments passed to f_dist, aside of the parameter.

SimulatedAnnealingABC.rho_mean — Method

Compute mean of distances WHY???

SimulatedAnnealingABC.sabc_noninf — Method

SABC_noninf(f_dist, d_prior, r_prior, n_sample, eps_init, iter_max, v, beta, delta, resample, verbose=1, adaptjump=true, summarystats=false, y, f_summarystats; p...)

SABC.noninf Algorithm

Arguments

f_dist: Function that either returns a random sample from the likelihood or the distance between data and such a random sample. The first argument must be the parameter vector.

d_prior: Function that returns the density of the prior distribution.

r_prior: Function which returns one vector of a random realization of the prior distribution.

n_sample: Desired number of samples.

eps_init: Initial epsilon.

iter_max: Maximal number of iterations.

v: Tuning parameter.

beta: Tuning parameter.

delta: Tuning parameter.

resample: After how many accepted updates?

p...: Further arguments passed to f_dist, aside of the parameter.

Returns

A named tuple with the following keys:

posterior_samples

prior_samples

eps: Value of epsilon at final iteration.

SimulatedAnnealingABC.schedule — Method

Solve for ϵ

See eq(31)