SimulatedAnnealingABC

Documentation for SimulatedAnnealingABC.

SimulatedAnnealingABC.estimate_jump_covariance

SimulatedAnnealingABC.initialization_noninf

SimulatedAnnealingABC.resample_ensample!

SimulatedAnnealingABC.rho_mean

SimulatedAnnealingABC.sabc

SimulatedAnnealingABC.schedule

SimulatedAnnealingABC.estimate_jump_covariance — Method

Estimate the coavariance for the jump distributions from an ensemble

SimulatedAnnealingABC.initialization_noninf — Method

Initialisation step

Arguments

See docs for sabc.

SimulatedAnnealingABC.resample_ensample! — Method

Resample ensample

SimulatedAnnealingABC.rho_mean — Method

Compute mean of distances WHY???

SimulatedAnnealingABC.sabc — Method

    sabc(f_dist,prior::Distribition, args...;
                     n_particles = 100, n_simulation=10_000,
                     eps_init = 1.0,
                     resample = n_particles,
                     v=0.3, β=1.0, δ=0.9,
                     adaptjump = true,
                     kwargs...)

Simulated Annealing Approximtaive Bayeisain Inference Algorithm

Arguments

f_dist: Function that distance between data and a random sample from the likelihood. The first argument must be the parameter vector.

prior: A Distribution defining the prior.

args...: Further arguments passed to f_dist

n_particles: Desired number of particles.

n_simulation: number of simulalations from f_dist.

eps_init: Initial epsilon.

v=0.3: Tuning parameter.

beta=1: Tuning parameter.

δ=0.9: Tuning parameter.

resample: After how many accepted updates?

kwargs...: Further arguments passed to f_dist`

Returns

A named tuple with the following keys:

posterior_samples

prior_samples

eps: value of epsilon at the last iteration.

SimulatedAnnealingABC.schedule — Method

Solve for ϵ

See eq(31)