API

SimulatedAnnealingABC.sabc — Function

sabc(f_dist::Function, prior::Distribution, args...;
      n_particles = 100, n_simulation = 10_000,
      algorithm = :single_eps,
      resample = 2*n_particles,
      v=1.0, β=0.8, δ=0.1,
      checkpoint_history = 1,
      show_progressbar::Bool = !is_logging(stderr),
      show_checkpoint = is_logging(stderr) ? 100 : Inf,
      kwargs...)

Simulated Annealing Approximate Bayesian Inference Algorithm

Arguments

f_dist: Function that returns one or more distances between data and a random sample from the likelihood. The first argument must be the parameter vector. See below for details.

prior: A Distribution defining the prior.

args...: Further arguments passed to f_dist

n_particles: Desired number of particles.

n_simulation: maximal number of simulations from f_dist.

v = 1.0: Tuning parameter for annealing speed. Must be positive.

β = 0.8: Tuning parameter for mixing. Between zero and one.

δ = 0.1: Tuning parameter for resampling intensity. Must be positive and should be small.

algorithm = :single_eps: Choose algorithm, either :multi_eps, or :single_eps. With :single_eps a global tolerance is used for all distances. Wit :multi_eps every distnace has it's own tolerance.

resample: After how many accepted population updates?

checkpoint_history = 1: every how many population updates distances and epsilons are stored

show_progressbar::Bool = !is_logging(stderr): defaults to true for interactive use.

show_checkpoint::Int = 100: every how many population updates algorithm state is displayed. By default disabled for for interactive use.

kwargs...: Further arguments passed to f_dist`

Details on f_dist

Given the observation $D$ and a stochastic model $f(θ)$ that provides a random sample from the likelihood $p(D|θ)$, the user provided function f_dist must be defined in one of the following ways depending on the algorithm used.

algorithm = :single_eps

In this case sabc expects that f_dist returns a positive scalar defined as:

f_dist(θ) = d(f(θ), D)

where $d()$ is a distance function. In practice often we want to compute this distance between summary statistics. If $s()$ computes one or multiple statistics of the data, f_dist becomes:

f_dist(θ) = d(s(f(θ)), s(D))

Note, in this situation it is up to the user to ensure that the distance function $d$ weights the different summary statistics meaningfully.

algorithm = :multi_eps

With the multi epsilon we can leave the weighting to the sampling algorithm. This introduces a seperate tolerance for each statistics. If we have K summary statistics, f_dist must return K distances:

f_dist(θ) = [ d_i(S_i(f(θ)), S_i(D)) for i in 1:k ]

Return

An object of type SABCresult

SimulatedAnnealingABC.update_population! — Function

update_population!(population_state::SABCresult,
                   f_dist, prior, args...;
                   n_simulation,
                   v=1.0, β=0.8, δ=0.1,
                   resample = 2*length(population_state.population),
                   checkpoint_history = 1,
                   show_progressbar::Bool = !is_logging(stderr),
                   show_checkpoint = is_logging(stderr) ? 100 : Inf,
                   kwargs...)

Updates particles with n_simulation and applies importance sampling if needed. Modifies population_state.

Arguments

See docstring for sabc.

SimulatedAnnealingABC.SABCresult — Type

Holds results from a SABC run with fields:

population: vector of parameter samples from the approximate posterior

u: transformed distances

ρ: distances

state: state of algorithm

The history of ϵ can be accessed with the field state.ϵ_history. The history of ρ can be accessed with the field state.ρ_history. The history of u can be accessed with the field state.u_history.